WS_liliang/requirement.txt
certifi==2019.3.9
chardet==3.0.4
ddt==1.2.1
dubbo-telnet==1.0.1b3
et-xmlfile==1.0.1
Faker==1.0.7
idna==2.8
jdcal==1.4.1
openpyxl==2.6.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-hessian==1.1.0
requests==2.21.0
selenium==3.141.0
six==1.12.0
suds-jurko==0.6
text-unidecode==1.2
urllib3==1.24.2
